mv test to test_dynamo.py
